Develop

+++ step2/database.py
@@ -1,15 +1,11 @@
-
-# employee = {
-#     {"login": "Alex", "password": "dontgooglegitler34", "salary": 60000, "date_raise": "20-03-2024"},
-#     {"login": "Bob", "password": "mommysony", "salary": 100, "date_raise": "22-01-2028"}
-# }
-
 from datetime import datetime
-from sqlalchemy.ext.asyncio import async_sessionmaker, create_async_engine
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
+from sqlalchemy.ext.asyncio import async_sessionmaker, create_async_engine
+
 
 class Model(DeclarativeBase):
     pass
+
 
 class TaskOrm(Model):
     __tablename__ = "tasks"
@@ -17,21 +13,20 @@ class TaskOrm(Model):
     name: Mapped[str]
     description: Mapped[str | None]
 
+
 class UserOrm(Model):
     __tablename__ = "users"
     id: Mapped[int] = mapped_column(primary_key=True)
     login: Mapped[str] 
     password: Mapped[str]
-    salary: Mapped[int]
+    salary: Mapped[float]
     next_raise_date: Mapped[str]
+
 
 engine = create_async_engine("sqlite+aiosqlite:///users.db")
 new_session = async_sessionmaker(engine, expire_on_commit=False)
 
 Base = DeclarativeBase()
-
-# Base.metadata.create_all(bind=engine)
-
 
 async def create_tables():
     async with engine.begin() as conn:
@@ -40,7 +35,3 @@ async def create_tables():
 async def delete_tables():
     async with engine.begin() as conn:
         await conn.run_sync(Model.metadata.drop_all)
-
-# Base.metadata.create_all(bind=engine)
-
-
